LayoutTests/http/tests/xmlhttprequest/redirect-credentials-responseURL-expected.txt

Test XMLHttpRequest responseURL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


req.readyState = 0: responseURL = 
PASS responseURL is ""
req.readyState = 1: responseURL = 
PASS responseURL is ""
req.readyState = 1: responseURL = 
PASS responseURL is ""
req.readyState = 2: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/basic-auth-default/dir1/basic-auth.php
req.readyState = 3: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/basic-auth-default/dir1/basic-auth.php
req.readyState = 4: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/basic-auth-default/dir1/basic-auth.php
PASS responseURL is "http://127.0.0.1:8000/xmlhttprequest/resources/basic-auth-default/dir1/basic-auth.php"
req.readyState = 0: responseURL = 
PASS responseURL is ""
req.readyState = 1: responseURL = 
PASS responseURL is ""
req.readyState = 1: responseURL = 
PASS responseURL is ""
req.readyState = 2: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/basic-auth-default/dir2/basic-auth.php
req.readyState = 3: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/basic-auth-default/dir2/basic-auth.php
req.readyState = 4: responseURL = http://127.0.0.1:8000/xmlhttprequest/resources/basic-auth-default/dir2/basic-auth.php
PASS responseURL is "http://127.0.0.1:8000/xmlhttprequest/resources/basic-auth-default/dir2/basic-auth.php"
req.readyState = 1: responseURL = 
PASS successfullyParsed is true

TEST COMPLETE
